pyproject.toml
[project]
name = "arcgeneral"
version = "0.1.0"
description = "LLM agent with stateful IPython REPL tool"
requires-python = ">=3.13"
dependencies = [
    "ipybox==0.6.2",
    "openrouter>=0.6.0",
    "python-dotenv>=1.2.1",
    "parallel-web>=0.3.4",
    "aiohttp>=3.9.0",
]

[project.scripts]
arcgeneral = "arcgeneral.cli:main"

[tool.setuptools.packages.find]
where = ["src/"]
include = ["arcgeneral*"]
